move renderTeacherPanel to teacherPanelHelpers
