+++ apps/src/code-studio/teacher.js
@@ -7,17 +7,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {getStore} from './redux';
-import ScriptTeacherPanel from './components/progress/ScriptTeacherPanel';
 import SectionSelector from './components/progress/SectionSelector';
 import ViewAsToggle from './components/progress/ViewAsToggle';
 import TeacherContentToggle from './components/TeacherContentToggle';
 import {ViewType, setViewType} from './viewAsRedux';
 import {lessonIsLockedForAllStudents} from '@cdo/apps/templates/progress/progressHelpers';
-import {setStudentsForCurrentSection} from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
 import {getHiddenStages} from './hiddenStageRedux';
 import commonMsg from '@cdo/locale';
-import {updateQueryParam} from '@cdo/apps/code-studio/utils';
-import {reload} from '@cdo/apps/utils';
 import {queryLockStatus} from './teacherPanelHelpers';
 
 function resizeScrollable() {
@@ -72,41 +68,6 @@ export function onReady() {
   });
 
   renderIntoLessonTeacherPanel();
-}
-
-/**
- * Render our teacher panel that shows up on our course overview page.
- */
-export function renderTeacherPanel(store, scriptId, section) {
-  const div = document.createElement('div');
-  div.setAttribute('id', 'teacher-panel-container');
-  queryLockStatus(store, scriptId);
-
-  if (section && section.students) {
-    store.dispatch(setStudentsForCurrentSection(section.id, section.students));
-  }
-
-  const onSelectUser = id => {
-    updateQueryParam('user_id', id);
-    reload();
-  };
-
-  const getSelectedUserId = () => {
-    const userIdStr = queryString.parse(location.search).user_id;
-    const selectedUserId = userIdStr ? parseInt(userIdStr, 10) : null;
-    return selectedUserId;
-  };
-
-  ReactDOM.render(
-    <Provider store={store}>
-      <ScriptTeacherPanel
-        onSelectUser={onSelectUser}
-        getSelectedUserId={getSelectedUserId}
-      />
-    </Provider>,
-    div
-  );
-  document.body.appendChild(div);
 }
 
 /**
